Update 2428(WRONG) - Capital.cpp

+++ uri-problems/Origins/OBI 2013/2428(WRONG) - Capital.cpp
@@ -5,12 +5,12 @@ using namespace std;
 
 int main() {
 
-    int a1, a2, a3, a4;
+    double a1, a2, a3, a4;
 
     cin >> a1 >> a2 >> a3 >> a4;
 
     double at = sqrt(a1) * sqrt(a2) * sqrt(a3) * sqrt(a4);
-    double dif = roundf(at * 100000000) / 100000000.0;
+    double dif = round(at * 100000000) / 100000000.0;
 
     if (dif - (int)dif <= 0.001) {
         cout << "S\n";
